perf: record densities as compact signed integers
DensityRecorder stored float64 by default. It now stores int16 for models
with volume exclusion (at most K particles per node and species) and int32
without, widening to int64 if a node ever exceeds that range, so no count
is truncated. Signed types keep differences between time points free of
unsigned wrap-around; NoVE identity-based timeevo no longer forces uint64.
The recording-budget estimate uses the resolved type.

+++ lgca/nove_ib_base.py
@@ -518,7 +518,7 @@ class NoVE_IBLGCA_base(NoVE_LGCA_base, IBLGCA_base, ABC):
         if recordN:
             observers.append(PopulationRecorder())
         if recorddens:
-            observers.append(DensityRecorder(dtype=np.uint))
+            observers.append(DensityRecorder())
         if recordchanneldens:
             observers.append(ChannelDensityRecorder())
         if recordfampop:
